+++ tests/Fixtures/Showcase/Shared/Model/User.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\Shared\Model;
+namespace App\Tests\Fixtures\Showcase\Shared\Model;
 
 use App\Engine\Model\Model;